[ABNF] Use markdown in doc comments, and a few fixes.
By using markdown in the documentation comments of the grammar, the markdown
file generated from the grammar includes those markdown features in the text,
making it more readable and better-looking.

Also fixed a few typos in the documentation comments.

Also updated a few documentation comments that were out of date after making
changes to the grammar.

Also removed a now-obsolete grammar rule for "input" parameters of functions.