[ABNF] Improve documentation and layout.
Explain better variable and constant declarations.

Leave one blank line between the rules for variable and constant declarations
(not necessary for ABNF, but consistent with the rest of the file and actually
expected by the ABNF-to-LaTeX converter).

Limit lines to 80 columns, by putting the rules for variable and constant
declarations over two lines each, with proper indentation.