[ABNF] Relax rule for end-of-line comments.
This was suggested by @bendyarm.

There is no need to require an end-of-line comment to terminate with an explicit
newline, which also prevents a file to end with and end-of-line comment without
a final newline. It suffices for an end-of-line comment to end when there is no
longer a non-newline character, which may happen either when there is a newline
character, or when the end of the file is reached.

This requires no change to the Leo lexer/parser, which is already implementing
the more relaxed rule.