[ABNF] Improve terminology.
This commit does not change the Leo language, but renames `package-name` to
`external-identifier`, at the same time establishing this nomenclature for
referring to these syntactic entities as 'external identifiers' instead of
'package names'.

Recall that these consist of lowercase letters, digits, and dashes, starting
with a letter, not ending with a dash, and with no two consecutive dashes.

The reason for the renaming is that these former package names, now external
identifiers, are used to name not only packages proper, but also directories and
files (the latter without the .leo extension): paths in import declarations use
these external identifiers to denote all these three kinds of entities
(packages, directories, and files). Recall that a Leo package is the whole
directory tree with the Leo.toml file: subdirectories of src/, and .leo files
under src/ or subdirectories, are not packages.

These external identifiers are also used to name package authors, in the
Leo.toml file. It's awkward to say that a package author is a package name; it's
cleaner to say that a package name is an external identifier, and that a package
author is also an external identifier.

In other words, this is just another kind of identifier, different from the ones
used to name entities declared within Leo files like functions and variables,
that is used to name things defined "outside" Leo files (packages, directories,
files, authors), which justifies the attribute 'external'.

It is a separate question whether we want Leo import declarations to use
identifiers instead of external identifiers. Even if we do that, this notion of
external identifier is still needed to describe the format of the Leo.toml file.