[ABNF] Replace package names with identifiers.
This is as discussed at today's sync.

The rule for package names is gone (but it will be re-introduced later to define
formally the format of certain TOML strings, not as part of the grammar of Leo
code, but as a separate grammar component).

Uses of package names in import declarations have been replaced with uses of
identifiers.

The changes to the grammar are fairly contained, as should be the changes to the
parser. Import resolution will have to translate between dashes and underscores.