[ABNF] Improve some rule names.
This does not change the language. It only renames two related rule names:

function-input => named-parameter
function-inputs => named-parameters

According to the new nomenclature, there are function parameters, which are self
parameters and named parameters: the nomenclature is clear and "symmetric".

The problem with the previous nomenclature is that both self and named
parameters are inputs, not just the named ones.