[ABNF] Re-classify some expressions in the grammar.
This does not change the language, but moves some expressions from postfix to
primary. Postfix expressions should have a recursive structure, which the moved
expressions do not have.

This change mirrors the organization of the parser of Leo written in ACL2, which
follows the naturally recursive structure of the language according to the
grammar.

Note that the current parser of Leo written in Rust, along with the Leo ASTs in
Rust, treats Leo functions as if they were expressions, but they are not
according to the ABNF grammar. (This is not necessarily a bug in the parser or
ASTs, if the expression ASTs are treated appropriately; but something we may
want to align to the ABNF grammar at some point -- separate topic).