[tests] Consistently use block comments.
There were four test files, found by @bendyarm, that use end-of-line comments
instead of block comments as prescribed by the README.md.

This commit changes them to block comments, for consistency with the README.md,
and uniformity across the files.

There is clearly nothing wrong with supporting block comments (which apparently
the Rust implementation of the Leo test framework already does). However, just
supporting block comments slightly simplifies the processing of the Leo test
files on the ACL2 side, as we are building that processing right now.

We could support end-of-line comments on the ACL2 side as well if desired. But
it seems unnecessary and unimportant at least for now. This is a quick PR, so
it's okay if it gets rejected and instead we want to support end-of-line
comments right away.