[tests] Fix format and expectation of a test.
This was apparently disabled, but it did not start with a block comment.

This commit fixes it to start with a block comment as required.

This commit also makes it a failing test, which is appropriate because in the
current version of Leo we do not allow untyped literals (i.e. lone numerals as
expressions), and it is thus a good negative test for the current version of
Leo. When we extend Leo to allow untyped literals, we can simply flip the
expectation of this test from Fail to Pass.