[tests] Move failing test to new fail.
Instead of commenting out this failing test, it seem better to add to a new file
for failing tests. If and when we change Leo to accept field literals in hex, we
can move the test from the Fail file to the Pass fail.

This also helps the ACL2 tester, which is currently not handling comment-out
tests of this form.