Refactor Pass trait and its implementations.
Now all that's public for each Pass implementation is a single struct,
with the visitor (if any) private to the module.

Made a new CompilerState struct that contains data of use to many
passes, such as the Ast, SymbolTable, etc, so that all passes have
access to those without having to be explicitly passed in and out.

Removed the individual functions like `static_analysis_pass` from the
Compiler; now each `do_pass` is just called directly.

Handler now contains an `Rc` and so implements Clone.