Write to array and struct members.
This is supported with a new pass, `WriteTransforming`,
which converts assignments to struct or array members
into assignments to variables, and any time the
struct or array is accessed, reconstructs it
from the variables.

Tests for this functionality are in `array_write.leo`.

Also fixed the test `assign_fail.leo`, which seems to have been
in a useless state for a long time.

Fixes #28389, #28430