Revise SourceMap and related types.
BytePos and CharPos and so on are gone - we just represent file offsets
with a `u32`.

SourceMap no longer tracks locations of multiple byte characters or
newlines -- this can just be handled when displaying data.

There's a new type `LineContents` that handles displaying spanned lines
to the user.

Lines and columns are both reported starting from index 1 -
older comments claimed to report columns from index 0, but this
wasn't true.

Fixes #28562.