Changes in preparation for next release.
Bump to snarkvm 4.0.0 and API changes.

Use "test" feature of snarkvm for execution tests.

Deploy twice in `run_with_ledger`. The double deployment is
temporarily necessary until upgrades are in place.

Introduce the option `--twice` in the CLI deploy command to
deploy twice and make local devnets useful.

Change the seed in primitive_casts.leo.
(this is a temporary workaround for ProvableHQ/snarkvm/2801)

Print outputs for `leo execute`. (This is not strictly necessary
for the release, but is a new feature that relies on new changes
to snarkvm's API.)

In execution, add programs twice.

Allow selecting more consensus versions.