Omit edition from the URL to fetch latest program with leo query program
As the title says. It looks like `/latest` endpoint has been deprecated
or was changed, but at the same time, if we omit it - we are able to
fetch the latest version.

So, this patch changes the current logic to omit `edition` in the URL if
user has not specified one, in order to fetch the latest one.