Improve shadowing error messages with multi-span support
Currently we only have information about a variable being shadowed and
the variable that it shadows it, but this is confusing for the user as
they don't see where is the variable that is being shadowed in the error
message.

So, this adds multi-span error reporting, as previous commits for
structs and records, reusing the same logic that was added.