Allow unlimited variable captures in async blocks
As the title says, but:

Currently, we have a limit that only 16 variables can be held in async
blocks.

So, this patch repairs it by creating "synthetic" structs when an async
block captures more than 16 variables, and also creates replacement
mappings for bundled variable accesses to refer to the struct directly
in the aleo bytecode.