Add support for empty arrays
This change adds support for empty array expressions and types and
generally improves handling of empty types in code generation.

Function arguments that resolve to empty types, such as `[u32;0]` are
ignored such that only non-empty arguments exist at runtime.
Empty return values are similarly ignored, or more accurately treated as
Unit is.

This support is currently restricted to empty arrays, but a future PR
could easily add support for `()` as an argument type or empty structs
with similar behavior.

Closes #28715