Unified intrinsics (#29034)
Implementation of the unified intrinsics proposal.

This change introduces intrinsic expressions, a special function invocation that start with _ and resolves to a compiler implemented behavior meant to replace the concept of a core function.

All associated constants, associated functions and special accesses are now converted by the parser into a corresponding intrinsic invocation.

This change is fully backwards compatible. The new syntax is additive and is equivalent to the old but no associated item or special access exists after parsing which simplifies expression handling in most of the compiler.

Closes #29010